+++ billing/__init__.py
@@ -1,0 +1,3 @@
+"""
+Billing package - Stripe metered billing integration
+"""